Larger terrain size (256), and blending using overlap regions
We are generating terrain patches that overlap by overlap units in both dimensions. After adding up all the overlapping patches in the terrain matrix, we divide by the count matrix to get the average value for each terrain point.

Still assumes that the GAN model can generate a square 2D patch of size 64x64 from a 100-dimensional random noise vector. The output of the GAN model is expected to be a single channel grayscale image where the pixel intensity corresponds to the height of the terrain at that point.